Use Parcelize for models serialization
