+++ sdk-core/src/main/java/com/reach5/identity/sdk/core/models/ProviderConfig.kt
@@ -1,45 +1,18 @@
 package com.reach5.identity.sdk.core.models
 
-import android.os.Parcel
 import android.os.Parcelable
+import kotlinx.android.parcel.Parcelize
 
+@Parcelize
 data class ProviderConfig(
     val provider: String,
     val clientId: String,
     val clientSecret: String?,
     val scope: Set<String>?
-) : Parcelable {
+) : Parcelable
 
-    constructor(parcel: Parcel) : this(
-        parcel.readString()!!,
-        parcel.readString()!!,
-        parcel.readString(),
-        parcel.createStringArrayList()?.toSet() ?: setOf<String>()
-    )
-
-    override fun writeToParcel(parcel: Parcel, flags: Int) {
-        parcel.writeString(provider)
-        parcel.writeString(clientId)
-        parcel.writeString(clientSecret)
-        parcel.writeStringList(scope?.toList())
-    }
-
-    override fun describeContents(): Int {
-        return 0
-    }
-
-    companion object CREATOR : Parcelable.Creator<ProviderConfig> {
-        override fun createFromParcel(parcel: Parcel): ProviderConfig {
-            return ProviderConfig(parcel)
-        }
-
-        override fun newArray(size: Int): Array<ProviderConfig?> {
-            return arrayOfNulls(size)
-        }
-    }
-}
-
+@Parcelize
 data class ProvidersConfigsResult(
     val items: Array<ProviderConfig>?,
     val status: String
-)
+) : Parcelable
