+++ sdk-core/src/main/java/com/reach5/identity/sdk/core/models/SdkInfos.kt
@@ -4,11 +4,11 @@ import android.os.Build
 
 object SdkInfos {
     const val version = "4.0.0"
-    const val platform = "android"
-    val device: String = Build.DEVICE
-    val model: String = Build.MODEL
-    val product: String = Build.PRODUCT
-    val apiLevel: Int = Build.VERSION.SDK_INT
+    private const val platform = "android"
+    private val device: String = Build.DEVICE
+    private val model: String = Build.MODEL
+    private val product: String = Build.PRODUCT
+    private val apiLevel: Int = Build.VERSION.SDK_INT
 
     private fun deviceInfo(): String {
         return "API Level $apiLevel, Device $device, Model $model, Product $product"
