+++ sdk-core/src/main/java/com/reach5/identity/sdk/core/models/UpdatePasswordRequest.kt
@@ -1,26 +1,35 @@
 package com.reach5.identity.sdk.core.models
 
-import com.google.gson.*
+import android.os.Parcelable
+import com.google.gson.Gson
+import com.google.gson.JsonElement
+import com.google.gson.JsonSerializationContext
+import com.google.gson.JsonSerializer
 import com.google.gson.annotations.SerializedName
+import kotlinx.android.parcel.Parcelize
 import java.lang.reflect.Type
 
 sealed class UpdatePasswordRequest {
+    @Parcelize
     data class FreshAccessTokenParams (
         val password: String
-    ) : UpdatePasswordRequest()
+    ) : UpdatePasswordRequest(), Parcelable
 
+    @Parcelize
     data class AccessTokenParams (
         @SerializedName("old_password")
         val oldPassword: String,
         val password: String
-    ) : UpdatePasswordRequest()
+    ) : UpdatePasswordRequest(), Parcelable
 
+    @Parcelize
     data class EmailParams (
         val email: String,
         val verificationCode: String,
         val password: String
-    ) : UpdatePasswordRequest()
+    ) : UpdatePasswordRequest(), Parcelable
 
+    @Parcelize
     data class EmailWithClientIdParams (
         val email: String,
         @SerializedName("verification_code")
@@ -28,14 +37,16 @@ sealed class UpdatePasswordRequest {
         val password: String,
         @SerializedName("client_id")
         val clientId: String
-    ) : UpdatePasswordRequest()
+    ) : UpdatePasswordRequest(), Parcelable
 
+    @Parcelize
     data class SmsParams (
         val phoneNumber: String,
         val verificationCode: String,
         val password: String
-    ) : UpdatePasswordRequest()
+    ) : UpdatePasswordRequest(), Parcelable
 
+    @Parcelize
     data class SmsWithClientIdParams (
         @SerializedName("phone_number")
         val phoneNumber: String,
@@ -44,7 +55,7 @@ sealed class UpdatePasswordRequest {
         val password: String,
         @SerializedName("client_id")
         val clientId: String
-    ) : UpdatePasswordRequest()
+    ) : UpdatePasswordRequest(), Parcelable
 
     companion object {
         fun<T : UpdatePasswordRequest> enrichWithClientId(params: T, clientId: String): UpdatePasswordRequest  {
